+++ src/main/java/org/mju_likelion/festival/term/controller/TermController.java
@@ -2,7 +2,7 @@ package org.mju_likelion.festival.term.controller;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
-import org.mju_likelion.festival.term.dto.TermResponse;
+import org.mju_likelion.festival.term.dto.response.TermResponse;
 import org.mju_likelion.festival.term.service.TermQueryService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
